Implement handling of IRF efficiency factors.
The energy dependent IRF efficiency factors are stored together with the effective area in a single response file. This information is now read by GLATAeff, which holds two functors of type GLATEfficiency that provide the energy dependent efficiency corrections. The efficiency information is optional. If no such information exists, the functors will be NULL.

The lifetime cube class GLATLtCube uses this information in the access operators that make use of effective area information. In that way, each detector section uses the proper efficiency factors. This is an improvement with respect to the ScienceTools which use an average efficiency correction for front and back together.