Initialise working array and apply new coding rules.
The working array values has not been initialised to 0 in the methods poisson_unbinned, poisson_binned and gaussian_binned. Although on most machine this initialization is apparently implicit, it cannot be guaranteed. The code now initialises explicitly the entire array to 0.

The code has also be overseen so that also one-line if statements or for loops are now in brackets. This enhances the readability of the code.