Normalize sky map internally and perform bi-linear skymap interpolation in eval and eval_gradients methods. Added filename constructor.
The sky map is now normalized internally, to that the total flux amounts to 1 ph/cm2/s. Note that this scaling can be modified using the normalization (or prefactor) parameter of the class, although in general this makes no sens as the flux is determined from the spectral law.

The eval and eval_gradients methods now make use of the sky direction access operator of the GSkymap class that performs bi-linear interpolation of the sky intensity based on the pixel values. This assures that eval and eval_gradients return a smooth function.

A filename constructor has been added that allows constructing a GModelSpatialMap object by loading a skymap from a FITS file.