Convergence problems no longer throw an exception.
For some reasons an empty exception was thrown in case of convergence problems by the romb() method. I guess this was done for debugging purposes, but this should not be the general behaviour of the method. I thus commented the throw directiv so that just warnings are issued but the code continues working.