Refactor CTA response class and implement support for diffuse source model.
The integrand classes that were part of the GCTAResponse class and that are used for computation of the response for extended sources was removed to a specific helper file. The classes do not longer belong to GCTAResponse, but are now standalone, and are locally defined in GCTAResponse_helpers.hpp. Some uniformisation of the code has been done.

The GCTAResponse::irf_diffuse and GCTAResponse::npred_diffuse methods have been implemented to support diffuse models based on spatial maps. The integration precision has been carefully adjusted to guarantee maximum speed while satisfying the precision requirements.