Add kluge to correct corrupt valgrind output.
The memcheck tool of valgrind apparently outputs too many </valgrindoutput> tags. This tag should only exist once at the end of the file. A kluge was added to the Makefile that strips all </valgrindoutput> tags from the file and adds a final </valgrindoutput> tag at the end of the file.