Add IAQ files from Werner Collmar:
u47569 is for a Phibar difference of 2 deg and a power-law on -2 over the 1-3 MeV energy range.
u47741 is for a Phibar difference of 2 deg and a power-law on -2 over the 3-10 MeV energy range.

Update also the Makefile for IAQ distribution and the unit test for checking of response loading.