Setup for development of COMPTEL module.
The directory tree has been setup for the development of the COMPTEL module. Instructions have been added to the various Makefile.am and to configure.ac to detect and to configure the library for inclusing of the COMPTEL module. The pyext/setup.py.in script has also bee adapted.

A simple test case has been written for loading of COMPTEL data into an observation.

The summary include file and the summary swig files have been setup.