Implement event bin, event cube and instrument direction.
Implemented the GCOMEventBin, GCOMEventCube and GCOMInstDir classes. An extensive unit test has been written to verify the functionnality of the classes. The event cube class successfully loads COMPTEL event data, the event bin class allows access to event information.

Note that a particularily clever implementation of GCOMEventBin has been made. This class does not hold members but pointers to members. There are two modes of operations: either the class allocates the members itself, and then also takes care about the deallocation, or it just holds the pointers from someone else, and thus does not deallocate the memory on destruction. This allows GCOMEventCube to use a single GCOMEventBin object for the manipulation of event data, while at the same time, GCOMEventBin can be used standalone for event manipulation.