Make compliant to coding standards and do minor fixes.
The code was overseen to comply to the GammaLib coding standards.

The m_newline member has been removed as this member is not used any longer.

The m_filename member is now initialized correctly in init_members() and copied in copy_members().

A new method named prefix() has been added that provides the prefix. This reduces code duplication in the append() method. Furthermore, a possible indent was put back that was removed during code revision.