Use forward declaration to avoid void pointer.
The class use void pointers to GSparseSymbolic and GSparseNumeric as the feature of forward declaration was not known when the class has been developed. This led to a number of complicated casts in the code. Instead of declaring void member pointers for the symbolic and numeric decomposition, GSparseSymbolic and GSparseNumeric pointers are now declared. All casts have been removed from the code.