Modify the Solaris kluge so that it only applies for the Sun Studio compiler.
We need a kluge to compile the swig wrappers using the Sun Studio compiler on Solaris. By default, the Sun Studio compiler tries to compile them with the C compiler, as the C compiler was also used to compile Python. swig wrapper files, however, are C++, hence they need a C++ compiler. We therefore declare a prefix variable PYTHON_BUILD_PREFIX that is used in the Makefile.am of the pyext directory. The characters in this variable will be prefixed to the python setup.py calls for building.

In case that a Sun Studio compiler is detected, we set the prefix to "CC=CC", which means that we set the Sun Studio C++ compiler (CC) to be the C compiler (environment variable CC).

Note that we also need an extra library Cstd for the Sun Studio compiler, that is defined by the PYTHON_EXTRA_LIBS parameter.