Make class non-abstract and implement Python testing.
As Python version < 3 does not support abstract base classes, we had to make GTestSuite abstract so that we can derive a test suite in Python.

A special append method was added to the Python interface that takes a PyObject as argument. The PyObject pointer is stored in a void array of pointers that is kept along with the function pointers. For a C++ function, the Python pointers are NULL. For a Python function, the C++ pointers are NULL. The run method checks then whether a C++ pointer or a Python pointer exists, and performs the adequat run command.