Add print_ methods to replace the dump_ methods.
In the old scheme we dumped the output directly in the output stream. Now we first print in a string, and then handle the string to whatever routine wants to use it. We thus added print_ methods that can remplace the old dump_ methods. Once all matrix classes use the print_ methods, the dump_ methods should be removed.