Limit maximum number of printed matrix rows and columns.
Added an integer parameter to print_elements() method that allows limiting the maximum number of rows and columns in the output. If more rows and/or columns are present than the number specified by the parameter, the excedent elements will be replaced by "..." in the centre of the matrix (i.e. only the corner elements of the matrix will be shown). If the parameter is set to 0, all matrix elements will be printed.

The parameter defaults to 10.