clang++/clang takes precedence over g++/gcc.
On new Mac OS X systems (>=10.7) we should select by default the clang++ compiler instead of the g++ compiler. The order of the compilers was thus changed on the compiler search, so that clang++ is detected before g++. As most other systems have no clang++ compiler (at least for the moment), this should not distrube the g++ discovery on these systems.