Search only for clang++ first if system is Mac OS X >= 10.7.
The compiler search order was changed to first check for clang. As this can lead to problems on system where an broken clang installation is present, we only want to search for clang first if we have a Mac OS X >= 10.7 system (this is the first system where Apple fully switched to clang).