Don't check for Python is Python bindings are disabled.
The configure script checked for Python, and added a Python include path for compilation even if Python bindings were not requested. Now, if Python bindings are disabled, no Python checks are performed. We still check for swig and the wrapper files for the record, but we also could disable this.