Use Darwin kluge definition to include gomp_thread_attr.
On Mac OS X using the Apple gcc compiler, we need to include gomp_thread_attr in case we want OpenMP support. Some code has been added to the configure script to detect this specific need, and only if this specific need is met the variable HAVE_OPENMP_DARWIN_KLUGE is defined in config.h. We use then this varaible to include (or not) the following sequence:
#ifdef HAVE_OPENMP_DARWIN_KLUGE
#include <pthread.h>
pthread_attr_t gomp_thread_attr;
#endif