Make compliant to MacPorts gcc compiler.
The Darwin kluge that consisted of linking statically with the GammaLib library for unit testing created a problem in the unit test when using the gcc compiler. Specifically, an exception thrown in purpose during the HealPix test was not caught.

It turned out that the problem was related to the static linking. Linking dynamically did not produce the problem. Furthermore, it also turned out that the Apple compiler also works using the dynamically linked library during unit testing. I have no idea why this was a problem before, but now it is no longer a problem.

I therefore removed the Darwin kluge of static linking.

Furthermore, a post-processing step was added that creates a symbolic link to the directory that contains the shared wrapper libraries. In that way, we do not need a double compilation anymore (before, we compiled once for the install and a second time for the unit test). This should now speed up the unit testing step.