Base classes on GRegistry interface.
All registry classes are now based on the GRegistry interface. The ostream and GLog friend functions have been removed from the classes as they are now implement on the GRegistry level.

The source code was checked to comply to the actual coding standards.

The Doxygen documentation of the source code was improved.