Rename CALDB to GAMMALIB_CALDB.
The environment variable CALDB may conflict with an existing Fermi CALDB environment variable. I therefore introduced a GAMMALIB_CALDB environment variable that takes precedence over CALDB. The GAMMALIB_CALDB environement variable will be used in the initialisation script, so that conflicts are solved. One can always change this environment variable later, or undefine it so that CALDB will be used.