Set m_npars in read_colnames().
As m_npars is used in the size() method, and the size() method is used in other reading methods, m_npars has to be as early as possible. We thus set m_npars at the end of the read_colnames() method, where the number of parameters is determined for the first time.