Make the load() and save() methods pure virtual and remove the read() and write() methods from the interface.
As the response files are not necessarily given in the FITS format (as example we can take the ASCII performance files), we remove the read() and write() methods from the abstract interface that take a GFits object as argument. Instead, we declare the load() and save() methods pure virtual, and we defer their implementation to the derived classes.