setup() method also accepts less than 2 nodes.
The previous change where the setup() method throws an error in case that less than 2 nodes are present was not viable. This does not allow successive addition of nodes to the node array. I thus reverted to the original code, skipping the setup step when less than 2 nodes are present.