Do not modify the Python wrapper files during the install step.
The Python wrapper files are modified by the script each time it is run. This changed the owners of these files during installation to root, making the ineditable afterwards. I changed the logic so that the wrapper file modifications are only done during the build step. No modifications are done during the install step, hence no file ownerships are changed anymore.