Remove post-processing of .py files and add list of classes and casts to __init__.py file.
The post-processing of .py files led to a modification of the .py files during the install process. I thus removed the post-processing step and added explicitely all classes and casts that should be imported. This is done by scanning all models for class and cast definitions, and by adding all names to the __all__ list.