Update cache after node access.
So far, the evaluation cache has not been updated after a node access. This led to a corrupt node array. A flag has been added that leads to cache update when node values have potentially changed.

I also declared cache members as mutable so that the setup and set_value methods can now be declared const. This makes the class cleaner.