Use the equivalent data type to allocate FITS image.
As the cfitsio routines perform automatic scaling of data, an image that is stored as long may need a floating point array to correctly hold the resulting values. The memory needs are correctly reflected by using the ffgiet routine from cfitsio, which returns the equivalent data type. This routine is now used.

This means however that a scaled integer image that is read from disk will be written back as floating point image.

See feature #610 for an action to modify the GFitsImage class.