Introduce support file and add com_wcs_mer2car() function.
The com_wcs_mer2car() function transforms the World Coordinate System of a sky map from Mercator's projection to a cartesian projection. This transformation is necessary to correctly interpret the COMPTEL data from the HEASARC site. The function only operates on Mercator's projections and does nothing for all other projections.