Remove radial() and dir() methods.
With the goal to adapt the extended source model interface to the other spatial model interfaces, the radial() and dir() methods have been removed. Instead of the radial() method, the spatial() method should be used. The dir() method is not really required as the GModelRadial class also has a dir() method.