Add observation identifier handling.
In addition to the applicability to a given instrument, models can now also be defined to be applicable to a given observation identifier. This allows for example to specify a separate background model for each observation. A model may also apply to several observation identifiers.

Methods id() have been added to convert from and to a string of observation identifiers.

The interface for the isvalid() method has been changed. An observation identifier argument has been added that allows to check for the applicability to an observation.