Add value cache and enhance bi-section algorithm.
A value cache has been added to that the interpolation computations are only done when the value or the nodes have been changed between subsequent class to @set_value()@ or @interpolate()@ (@interpolate()@ calls in fact @set_value()@, so that the changes have only be done in the @set_value()@ method).

An unnecessary if statement has been removed from the bi-section algorithm.

See feature #102.