Add parameter status determination and isvalid(), isnotanumber() and isvalid() methods.
Determines the parameter status in the set_value() method.

The special values INDEF, NONE, UNDEF and UNDEFINED (case insensitive) can now be passed to integer and floating point parameters, leading to a status of "undefined". This status can be checked using the isundefined() method.

The spacial values INF, INFINITY or NAN (case insensitive) can now be passed to integer and floating point parameters. For integer parameter, the maximum long number will be assigned. For floating point parameter, the status will be changed to "not a number". This status can be checked using the isnotanumber() method.

All other values will lead to a valid status, which can be checked using the isvalid() method.