Adapt to new GTime interface and handle time reference.
The class has been adapted to the new GTime interface.

In addition, it now stores the time reference that is used when reading and writing data from and to a FITS file. In that way, the original time reference system is preserved, while times are always stored in the native GammaLib time reference.