Modify GTime interface.
The GTime class now represents the time in a GammaLib native time reference. The reference is no longer stored as a class member.

Conversion methods have been added to convert the time into any possible reference system. The time reference systems is represented by the new class GTimeReference.

The met() method has been removed, and replaced by the secs() method which returns the time in the native format in units of seconds. This is the way how GTime stores the time. An additional method days() gives the time in the same reference in units of days.