By default, write MJD reference in 2 keywords.
Write the MJD reference into 2 keywords, splitting into the integer and the fractional part, assures the highest accuracy for the reference time. We use this format now as the default, and only write in a single floating point value if the FITS file contains already a single floating point value keyword.