Add support for instrument specific model scale factors.
A vector of GModelPar objects has been added to the GModel class that can hold instrument specific model scaling factors. The name of each parameter specifies the instrument, which will be converted automatically in upper case.

The scale factors can be set and accessed using the scale() methods. If the scale method is called for an instrument for which no instrument specific scale factor has been defined, unity is returned (hence the model will not be scaled).

The class now provides the protected methods read_scales() and write_scales() for reading and writing of scale factors from and to a XML file. The scale factors are written to an optional <scaling> tag, which contains a list of tags of type <instrument ... />. These methods are used by the GModelSky class for reading and writing the scale factors.