Make class non-abstract.
The GModelSky class is no longer abstract but is the general implementation of a sky model. For this purpose, the member m_type has been added that holds a string specifying the model type.

The type is now set automatically based on the type of the spatial component using the protected set_type() method.

More constructors have been added to allow building an object from the model components, including also the temporal component. Also, another constructor was added to allow using also temporal information from an XML element for construction.