The GModelSky class is no longer abstract but is the general implementation of a sky model.
The type is now set automatically based on the type of the spatial component using the protected set_type() method. The set_type() method is called by all non-void constructors and the XML read method.

The type provided in the XML file is for the moment ignored. The type will be set directly from the type of the spatial component.

More constructors have been added to allow building an object from the model components, including also the temporal component. Also, another constructor was added to allow using also temporal information from an XML element for construction.

Three instances of GModelSky are now registered using the types PointSource, ExtendedSource and DiffuseSource. This is only needed for correct interpretation of XML files.