Allow insertion in empty object and avoid endless loop for self-extension.
A special treatment was added for insertion of elements in empty objects. This was not possible before because any index was outside the valid range for an empty object.

The number of elements is now set before appending a container to a container, so that self-extension does not lead to an endless loop. Before, if a container was self-extended, the size parameter grew with the extension, leading to a situation where the loop never ended.