Review code before integration.
I quickly went over the code before integration to make it compliant to the coding standards. Most of the modifications were cosmetical (related to code formating)

In GModelSpatialEllipticalDisk::eval() I removed the std::pow(..., 2) calls and replaced them by simple multiplications as the later are much faster. This should speed up the computation somewhat. More  importantly, I added an initial check on theta to see whether the position is located within the enclosing circle of the ellipse. This will simply return 0 for any locations outside the circle, avoiding unnecessary computations.

In GModelSpatialEllipticalDisk::theta_max(), I added a check on whether major() or minor() are the larger value, and based on the check, I return the larger of both. This is important as a model fit may make minor() larger than major().

In GModelSpatialEllipticalDisk::read() and GModelSpatialEllipticalDisk::write(), a few bugs have been corrected.