Implement parsing of text string.
The parse() method now either parses a file or a text string, depending on whether a valid file pointer or text string is provided. This option is now used in the GXml(const std::string& xml) constructor that will interpret the argument as either a text string or a file name. The interpretation is done on basis of the 5 first characters. If the string starts with "<?xml" the string is an XML document, otherwise it will be interpreted as a filename.