Add m_name, m_host and m_port members and implement new socket handling.
The Hub host and port strings are now stored in the members m_host and m_port. This prevents recomuting them when connecting to the Hub. Also the client name is now stored. It will be sent as the minimum required metadata (although it is formally not required).

The socket handling now follows the recommandations of Beej's Guide to Network Programming.