Use GUrl class instead of FILE type for reading and writing.
The use of the new GUrl class insread of the FILE type allows now for transparent input and output of an XML document in different URL types. So far only the file type URL exists (GUrlFile), but this can now easily be expanded to support piping into a string or direct http transfer. The latter is particularily useful for XML-RPC implementation.