Correct optimization bug.
The GObservations::copy_members() copied the m_fct which it should not have done, as the m_fct::m_this member shall always point to GObservations, hence it should never be copied! Copying led in fact to a segmentation fault. The copy instruction has been removed which corrected the bug.