Declare methods inline.
The methods that are implemented in the header file are now explicitely declared inline and put in a dedicated section after the class declaration. This enables proper documentation of the methods.

A small semantic bug was also fixed in the reserve() method where the reserve class was put after return. This apparently worked, but is not correct as the method returns void.