Rework interface of eval, eval_gradients and mc methods.
The eval() and eval_gradients() methods now take a GPhoton object as argument to allow the implementation of energy and time dependent source morphologies.

GEnergy and GTime objects have been added to the mc() method interface to support equally the implementation of energy and time dependent source morphologies.

So far, no class really implements energy and time dependent source morphologies, but now we have the infrastructure to do so.