Complete value constructor, add member access methods and introduce precomputation cache.
The value constructor now takes values for all 4 parameter members as arguments. The pivot and cut-off energy arguments have been declared as GEnergy.

All parameter members now have set and retrieve methods. The pivot and cut-off method set and retrieve GEnergy objects.

A precomputation cache has been added for the eval() and eval_gradients() methods that does accelerate computations in case that the energy and the index, cutoff and pivot energy parameters have not changed. This should improve the performance for binned analysis.