Modify value constructor, add methods for parameter access, add evaluation and mc cache.
The value constructor now takes an object of type GEnergy for the pivot energy.

Methods have been added to access are three model parameters.

Evaluation caches have been added to support the eval(), eval_gradients() and mc() methods. This speeds up evaluation of the spectrum, in particular for binned analysis, and the simulation of photons.