The GSource value constructor takes now a GModelSpatial pointer instead of a const reference, and the model is no longer cloned.
Model cloning is time consuming as it implies dynamic memory allocation, and the GSource class is used for very low-level functions in the model computation, so we don't want to have a speed penalty here. Therefore, GSource now carries the model information instead of cloning it. If cloned information is needed, one can alway pass a cloned pointer to GSource.