Review and modify abstract matrix interface.
The abstract matrix interface defined in GMatrixBase was reviewed and modified according to common practices.

All "col" arguments and method names were expanded to "column". The cols() method which returned the number of columns in a matrix is now named columns().

The insert_row/col and set_row/col methods were renamed to row() and column() which allow setting and getting a row or a column in a matrix using a GVector object. In the set methods, the row or column index are now passed as first argument, followed by the vector.

The add_row/col methods were renamed to add_to_row() and add_to_column() to be more explicit. Also here, the row or column index are now passed as first argument, followed by the vector.

Every matrix class will now have a transpose(), invert(), negate(), abs(), fill(), min(), max(), and sum() method.

A size() method has been added to GMatrixBase that returns the number of elements in a matrix.