Assure unique model names.
The set(), append(), insert(), and extend() methods now verify that the added models have unique names, i.e. that the name of a model that is added does not exist already in the model container. In that way, all models in a model container *should* have unique names.

Note, however, that a client can still violate this rule by retrieving a pointer to a model and by changing the name of the model explicitly.