Declare index access operators as inline, rename hasmodel() to contains() and add Doxygen documentation for class with examples.
The index access operators [] have now be placed inline so that faster code can be generated (as the index operator does not index checking it is now pretty short code).

The hasmodel() methods should semantically be named contains() (in general, a container should always have the contains() method to check for names).

The Doxygen documentation of the class has been enhanced, adding usage example to the general description.