Add contains() and at() methods, declare some methods inline, and control uniqueness of observation identifiers for a given instrument.
The contains() method has been added that checks whether an observation for a given instrument and a given identifier exists.

The operator [] does not check the index range anymore (the operators are now declared inline). For index check, the at() methods have been added.

Methods that were implemented in the header file are now declared inline.

The set(), append(), insert() and extend() methods now verify that the observation that should be added has an identifier that does not yet exist for the same instrument in the container. In other words, observation identifiers need to be unique for a given instrument.