Add filename() method.
The filename() method returns the file name from which information has been loaded or saved. A mutable m_filename member has been created for this that stores the file name in the GArf, GPha and GRmf classes. Tests have been added in the respective test suite to verify the correct functioning of the methods.