Rework map cube spatial model class.
The map cube is no longer normalised, but the values are taken as given. It is assumed that the map cube pixels are given in units of ph/cm2/s/sr/MeV.

The eval() and eval_gradients() methods now provide interpolation of the maps in energy. For this purpose, the energy values of the maps are stored in log10(energy) in the m_logE member node array.

The mc() method has now also been implemented. It takes the nearest map from the map cube and returns a random sky direction following the intensity distribution of this map.

A spectrum() method has been implemented that returns the spatially integrated intensity of each map in the map cube as a GModelSpectralNodes object. This method is used in GModelSky for Monte Carlo simulations.