Implement diffuse map cube support in mc() method.
The mc() method now explicitely support Monte Carlo simulation for diffuse map cubes. The method detects whether a spatial component is of type GModelSpatialDiffuseCube. In case, it takes the diffuse map cube spectrum as the spectral component, implemented by a GModelSpectralNodes object. Any existing spectral component is multiplied to the nodes model by evaluation the spectral function at the node energies.

Note that the method still has a limitation in that the simulation cone is not taken into account when extracting the diffuse map cube spectrum.