Make energy boundary loading and saving optional.
The XSPEC PHA format does not store the energy boundaries together with the PHA file. To support the original XSPEC PHA format, loading and saving of energy boundaries is now optional. A constructor has been added that allows constructing of a GPha object by specifying the number of channels only.