Correctly handle saving of string columns and declare all methods implemented in the header file as inline methods.
During the rework, a bug occured in the string column class that prevented from correct saving to the file. The bug was due to the fact that the wrong pointer was specified by the ptr_data() method, and then, also the GFitsTableCol::isloaded() method that worked on ptr_data() was not corrected. The ptr_data() method has been corrected for the string column class and the isloaded() method has been defined as pure virtual in GFitsTableCol, and now is implemented in the derived classes. We thus have full control over the definition of the loading state.

To conform to the actual coding standards, all methods that are implemented by the table column classes in the header file are now declared inline.