Correctly read CTA event lists that do not conform to standard.
So far, the CTA event list reader explicitly casted the column pointers to the expected format, which led to problems when the format was not the expected one (e.g. double instead of float). In fact, the code casted in this case a GFitsTableDoubleCol pointer to a GFitsTableFloatCol pointer, and the returned values were corrupted.

Now the code uses a generic pointer to GFitsTableCol and uses the real() and integer() method to access their contents. This could lead to some overflow in case that the FITS value does not fit in the range.