Introduce --enable-warnings option.
The configure option --enable-warnings now sets the definition G_WARNINGS in the config.h file.

The function gammalib::warning() has been added to GTools. The function dumps a warning to the console in a standard format if the warning option is enabled.

The function is now used in GFitsHeaderCard to dump any warning about invalid header card value types to the screen.