Catch pixel-to-spherical transformation exceptions in GCTAEventCube::set_directions.
When using the Hammer-Aitoff projection, the pixel-to-spherical used in GCTAEventCube::set_directions may fell as pixels may correspond to zones that do not represent sky regions. To prevent this problem, any exception is catched and the solid angle of the pixel is set to 0. This has the nice side effect of filtering out the bin during model fitting as the solid angle is used to multiply the model value.