Properly deleted empty FITS file.
If a FITS file is created using the open() method and if HDUs have been appended to this FITS file, the GFits::close() method (or the destructor) resulted in an exception due to an invalid filepointer. The result FITS file was empty, with a single "END" statement.

Now, the m_created flag is used to signal that a file has been created but not yet saved. Calling the GFits::save() method will set the flag to false. The GFits::free_member() method will delete the FITS file if m_created is still true.