Review GFitsHDU interface and remove fits_key_not_found exception.
The GFitsHDU::card methods now return a reference instead of a pointer. Some more methods have been added to GFitsHDU for header card access and management.

The fits_key_not_found exception has been removed, and keyword checks are now done using the hascard() methods. This was needed as the GFitsHeader class does not throw the fits_key_not_found exception anymore if a header card is not found.

The FITS utility function gammalib::fits_move_to_hdu has been added to the GFits.hpp/cpp files which provides support for moving to a specific HDU. As this cfitsio function is widely used in the FITS module it was favourable to put it in a specific function with a generic exception message.